Force data validation and html injeciton fixes
major problem is that it forces a try{} catch{} for filenames when
adding to prevent it from throwing errors in usage. Outweighs cons
though.

Also added a quick method in the add command that prevents duplicates
(not existant in the insert command)

Also used the webutility class to prevent encode data in the webserver
class. Should prevent injection now