Naming fixes and Self-Defining Comparisons
Commands now can override and implement their own method to see if the
command is suitable, this has fixed a lot of decoupling and can be
further implemented. As compensation a "get command" method was added.

The mapmodulelist thing will contain both the full map list and also the
blacklist one

Renamed MessageData for better readability and added a destination
parameter